Harden savings platform: exact money, auditable ledger, security fixes
Brings the platform from a prototype to something that can be operated:
correct decimal arithmetic, an append-only ledger, working authorization,
a supported framework, and non-custodial mobile-money ingestion.

Verified: ./mvnw clean verify -> 61/61 tests, coverage 25% -> 31%.
Frontend lint and build both exit 0 (neither had ever been run before).
All 8 migrations apply twice against PostgreSQL 18 with zero failures, and
ddl-auto=validate boots clean against the resulting schema.

MONEY CORRECTNESS

* Convert all 34 monetary fields from Double to BigDecimal with explicit
  NUMERIC(19,2) columns (interest rate NUMERIC(9,4)). Binary floating point
  cannot represent decimal currency, so group totals drifted from the sum of
  member balances. Demonstrated in Postgres: SUM(0.10+0.20+1000.55) returned
  1000.8499999999999 before, 1000.85 after.
* Add Money with fixed scale and HALF_UP, rounding only at boundaries.
* Add an append-only ledger. Balances are derived by summing rows, never
  stored, so any balance can be explained. Corrections append a reversing
  entry; UPDATE and DELETE are refused by database triggers.
* Make money writes idempotent on a caller-supplied key, so a retry on a poor
  connection cannot double-count a contribution.
* Fix Savings.type persisting as an enum ORDINAL: reordering the enum would
  have silently reinterpreted every historical savings row.
* Add @Version optimistic locking to SavingsCycle and MemberPayout.
* Fix the savings ledger overwriting instead of summing same-day entries.

CRITICAL PRE-EXISTING BUGS

* User and SavingsGroup generated equals/hashCode across their bidirectional
  association and recursed into each other, so any HashSet insert threw
  StackOverflowError. Group creation had never worked at runtime.
* The super admin could never log in: login required a savingsGroupId and
  verified group membership, but the super admin belongs to no group.
* The JWT identity chain was broken end to end: tokens carried the username as
  subject while every lookup was by email, so the principal could not be
  resolved on any request after login.
* AuthenticationManager was derived from the shared HttpSecurity builder and
  had no usable provider, failing every login.
* LoansPage called setLoans(), which does not exist, so Approve, Reject and
  Delete all threw ReferenceError and faked their API calls with setTimeout.
* The mobile drawer backdrop rendered while the sidebar was closed and its
  click handler opened it, covering the whole app on a phone.
* apiSlice and the feature API modules defined duplicate endpoint names; RTK
  Query keeps the first registration, so broken paths shadowed correct ones.
* The Vite dev proxy stripped the /api prefix that every controller is mapped
  under, returning 404 for every request.
* GroupService.createGroup dereferenced the null admin the controller passes.
* GroupServiceTest never compiled, so the suite had never run.

SECURITY

* Fix all 41 @PreAuthorize expressions. They were wrong in two directions:
  hasRole('ADMIN') referenced a role that does not exist, and
  hasRole('ROLE_SUPER_ADMIN') resolved to ROLE_ROLE_SUPER_ADMIN. Admin-only
  endpoints were unreachable by anyone.
* Add object-level authorization to every member-scoped endpoint. Any
  authenticated member could previously read another member's savings, loans,
  fines and reports by changing the id in the URL.
* Fix two cross-tenant leaks: GET /api/users/active returned every active user
  in the system to any group admin, and /api/users/{id} plus
  /api/users/{id}/status were role-gated only.
* Require super admin for group creation, which provisions a ROLE_GROUP_ADMIN
  account and was previously unauthenticated.
* Move every secret to the environment; the app refuses to start without them.
  Stop logging generated passwords, and return the group admin's one-time
  password to the caller instead.
* Remove the duplicate JWT filter, add JSON 401/403 handlers, and replace
  placeholder getCurrentUserId() returning 1L with the real principal.
* Add validation constraints and a global exception handler returning RFC 9457
  problem details; @Valid was previously a no-op and 49 raw RuntimeExceptions
  became 400s that leaked internal messages.
* Wire the AuditLog entity, which was read but never written to.

SCHEMA AND BUILD

* Add V1__init.sql, generated from the entities, and set ddl-auto=validate.
  Migrations previously started at V2 while Hibernate invented the schema.
* Remove a hardcoded bcrypt hash and an invalid role value from V2. See the
  upgrade note below.
* Consolidate on Maven (Java 21) and add the wrapper; the Gradle build
  targeted a different toolchain.
* Upgrade Spring Boot 3.2.3 -> 4.1.1 (3.2 lost OSS support in Nov 2024), with
  Spring Security 7, jjwt 0.12 and springdoc 3.x API migrations.
* Untrack 86 committed .class files and add gitignores.

OPERATIONS

* Add Actuator with health public and metrics super-admin only, dev/prod
  profiles, JSON logging with credential masking, and request correlation ids.
* Add Dockerfiles (non-root, JRE-only), compose with health gating and
  read-only filesystems, CI that proves migrations are idempotent, and
  Dependabot.
* Paginate the endpoints that grew without bound and cap page size.
* Split the frontend bundle: entry chunk 1,070 kB -> 154 kB (327 -> 46 kB
  gzip), with xlsx no longer downloaded to reach the login screen.

MOBILE MONEY (record-keeping only)

* Add a provider-agnostic ingestion pipeline. The platform never holds member
  funds, which is what keeps it outside BNR PSP and e-money licensing; the
  constraint is documented and enforced in the package.
* Webhooks are HMAC-verified over exact bytes with constant-time comparison
  and a replay window; an unconfigured secret refuses everything.
* Idempotent on the provider's event id, so a provider retry cannot
  double-credit. Unmatched or ambiguous payers are parked for a human rather
  than guessed at.
* Add reconciliation that reports provider-versus-ledger drift without
  auto-correcting it.

PRIVACY (Law 058/2021)

* Add per-purpose versioned consent, append-only.
* Add subject-access export and erasure. Erasure anonymises personal data and
  retains financial records, so the right to be forgotten and the duty to keep
  books can both be honoured without corrupting group totals.

FRONTEND

* Add an installable PWA with an offline contribution queue that replays
  through the idempotency keys above, cleared on sign-out because handsets are
  shared. API responses are never cached.
* Replace the 758-line hand-rolled translation context with i18next, with
  locale-aware RWF and date formatting.
* Resolve 5,963 lint errors to zero and surface query errors that were
  previously destructured and dropped.

UPGRADE NOTE

Editing V2 changed its Flyway checksum. Any database that already ran the old
V2 needs `flyway repair` once before the next migrate.

The JWT secret previously committed in application.properties is compromised
and remains in git history. It has been replaced by an environment variable;
generate a new value and do not reuse the old one.

KNOWN GAPS

Coverage is 31%: service 13%, controller 7%, dto 0%. Tailwind 3->4 and
React 18->19 not attempted. xlsx@0.18.5 carries unpatched advisories and is
the newest version on npm. SMS/USSD and outbound disbursement are not built.
NCSA registration, a DPO, data localisation and a Kinyarwanda privacy notice
are organisational obligations; see docs/PRIVACY.md.

+++ ikimina/src/main/java/com/example/ikimina/dto/MemberPayoutDTO.java
@@ -1,5 +1,7 @@
 package com.example.ikimina.dto;
 
+import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -10,9 +12,9 @@ public class MemberPayoutDTO {
     private Long memberId;
     private String memberName;
     private String memberNumber;
-    private Double totalUbwizigame;
-    private Double totalIngoboka;
-    private Double payoutAmount;
+    private BigDecimal totalUbwizigame;
+    private BigDecimal totalIngoboka;
+    private BigDecimal payoutAmount;
     private LocalDate createdAt;
     private LocalDate paidAt;
     private String status;
