+++ ikimina/src/main/java/com/example/ikimina/dto/SavingsCycleDTO.java
@@ -1,5 +1,7 @@
 package com.example.ikimina.dto;
 
+import java.math.BigDecimal;
+
 import java.time.LocalDate;
 import java.util.List;
 
@@ -13,9 +15,9 @@ public class SavingsCycleDTO {
     private LocalDate startDate;
     private LocalDate endDate;
     private String status;
-    private Double totalUbwizigameCollected;
-    private Double totalIngobokaCollected;
-    private Double totalUbwizigameDistributed;
+    private BigDecimal totalUbwizigameCollected;
+    private BigDecimal totalIngobokaCollected;
+    private BigDecimal totalUbwizigameDistributed;
     private LocalDate createdAt;
     private LocalDate distributedAt;
     private List<MemberPayoutDTO> memberPayouts;
